+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - `summary-details` - Add new `autofocus` option to allow the first focusable element inside to receive focus automatically.
 - `summary-details` - Now automatically opens if the browser's search function finds matching content within.
 - `summary-details` - New `toggletip` option announces content automatically; useful for "info" tooltips, for example.
+- `accordion-group` - Now has a "close all panels" button, as well as an "open all panels".
+- `accordion-group` - Will now open panels if the browser's search function finds matching content within.
 
 ### Fixes